Refactor: reducer_actions를 api_actions과 _actions로 분리
